+++ src/components/Questionnaire.jsx
@@ -3,74 +3,74 @@ import {db} from '../../config/firestore'
 import { collection, addDoc, serverTimestamp} from 'firebase/firestore'
 import { getAuth } from 'firebase/auth'
 
-export default function Questionnaire(){
+export default function Questionnaire(props){
 
-    const [formData, setFormData] = React.useState({
-        q1:null,
-        q2:null,
-    })
+    // const [formData, setFormData] = React.useState({
+    //     q1:null,
+    //     q2:null,
+    // })
 
-    const [calculatedMetrics, setCalculatedMetrics] = React.useState({
-        totalScore: null
-    })
+    // const [calculatedMetrics, setCalculatedMetrics] = React.useState({
+    //     totalScore: null
+    // })
 
-    React.useEffect(() => {
-        calculateMetrics()
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [formData])
+    // React.useEffect(() => {
+    //     calculateMetrics()
+    //     // eslint-disable-next-line react-hooks/exhaustive-deps
+    // }, [formData])
 
-    function handleChange(e){
-        const {name, value} = e.target
-        setFormData(prevFormData => {
-            return {
-                ...prevFormData,
-                [name]:Number(value)
-            }
-        })
-    }
+    // function handleChange(e){
+    //     const {name, value} = e.target
+    //     setFormData(prevFormData => {
+    //         return {
+    //             ...prevFormData,
+    //             [name]:Number(value)
+    //         }
+    //     })
+    // }
 
-    function calculateMetrics(){
-        let sum = 0
-        for( let key in formData){
-            sum += formData[key]
-        }
-        setCalculatedMetrics( prevCalculatedMetrics => {
-            return {
-                ...prevCalculatedMetrics,
-                totalScore: sum
-            }
-        })
-    }
+    // function calculateMetrics(){
+    //     let sum = 0
+    //     for( let key in formData){
+    //         sum += formData[key]
+    //     }
+    //     setCalculatedMetrics( prevCalculatedMetrics => {
+    //         return {
+    //             ...prevCalculatedMetrics,
+    //             totalScore: sum
+    //         }
+    //     })
+    // }
 
-    async function addResult(){
-        const auth = getAuth()
-        const user = auth.currentUser
+    // async function addResult(){
+    //     const auth = getAuth()
+    //     const user = auth.currentUser
         
-        const currentTime = serverTimestamp()
+    //     const currentTime = serverTimestamp()
 
-        const newDocValues = {
-            createdBy: user.uid,
-            createdAt: currentTime,
-            ...formData,
-            ...calculatedMetrics,
-        }
-        await addDoc(collection(db, "HABC-responses"), newDocValues);
-    }
+    //     const newDocValues = {
+    //         createdBy: user.uid,
+    //         createdAt: currentTime,
+    //         ...formData,
+    //         ...calculatedMetrics,
+    //     }
+    //     await addDoc(collection(db, "HABC-responses"), newDocValues);
+    // }
 
-    function handleSubmit(e){
-        e.preventDefault()
-        try{
-            addResult()
-            console.log('added doc to firestore')
-        }catch(error){
-            console.error(error)
-        }
-    } 
+    // function handleSubmit(e){
+    //     e.preventDefault()
+    //     try{
+    //         addResult()
+    //         console.log('added doc to firestore')
+    //     }catch(error){
+    //         console.error(error)
+    //     }
+    // } 
 
     return(
         <>
             <h2>Fill out and submit this form please</h2>
-            <form onSubmit={handleSubmit}>
+            <form onSubmit={props.handleSubmit}>
                 <fieldset>
                     <legend>Question 1</legend>
                     <label >
@@ -78,8 +78,8 @@ export default function Questionnaire(){
                             type="radio"
                             name='q1'
                             value={0}
-                            checked={formData.q1 === 0}
-                            onChange={handleChange}
+                            checked={props.formData.q1 === 0}
+                            onChange={props.handleChange}
                          />
                         None
                     </label>
@@ -88,8 +88,8 @@ export default function Questionnaire(){
                             type="radio"
                             name='q1'
                             value={1}
-                            checked={formData.q1 === 1}
-                            onChange={handleChange}
+                            checked={props.formData.q1 === 1}
+                            onChange={props.handleChange}
                          />
                         Some
                     </label>
@@ -98,8 +98,8 @@ export default function Questionnaire(){
                             type="radio"
                             name='q1'
                             value={2}
-                            checked={formData.q1 === 2}
-                            onChange={handleChange}
+                            checked={props.formData.q1 === 2}
+                            onChange={props.handleChange}
                          />
                         Often
                     </label>
@@ -108,8 +108,8 @@ export default function Questionnaire(){
                             type="radio"
                             name='q1'
                             value={3}
-                            checked={formData.q1 === 3}
-                            onChange={handleChange}
+                            checked={props.formData.q1 === 3}
+                            onChange={props.handleChange}
                          />
                         Always
                     </label>
@@ -121,8 +121,8 @@ export default function Questionnaire(){
                             type="radio"
                             name='q2'
                             value={0}
-                            checked={formData.q2 === 0}
-                            onChange={handleChange}
+                            checked={props.formData.q2 === 0}
+                            onChange={props.handleChange}
                          />
                         None
                     </label>
@@ -131,8 +131,8 @@ export default function Questionnaire(){
                             type="radio"
                             name='q2'
                             value={1}
-                            checked={formData.q2 === 1}
-                            onChange={handleChange}
+                            checked={props.formData.q2 === 1}
+                            onChange={props.handleChange}
                          />
                         Some
                     </label>
@@ -141,8 +141,8 @@ export default function Questionnaire(){
                             type="radio"
                             name='q2'
                             value={2}
-                            checked={formData.q2 === 2}
-                            onChange={handleChange}
+                            checked={props.formData.q2 === 2}
+                            onChange={props.handleChange}
                          />
                         Often
                     </label>
@@ -151,8 +151,8 @@ export default function Questionnaire(){
                             type="radio"
                             name='q2'
                             value={3}
-                            checked={formData.q2 === 3}
-                            onChange={handleChange}
+                            checked={props.formData.q2 === 3}
+                            onChange={props.handleChange}
                          />
                         Always
                     </label>
