+++ src/components/Results.jsx
@@ -5,21 +5,21 @@ import {getAuth} from 'firebase/auth'
 
 export default function Results(){
 
-    async function getResults(){
-        const auth = getAuth()
-        const user = auth.currentUser
-        console.log(user.uid)
+    // async function getResults(){
+    //     const auth = getAuth()
+    //     const user = auth.currentUser
+    //     console.log(user.uid)
 
-        const userDocs = []
-        const q = query(collection(db, 'HABC-responses'), where('createdBy', '==', `${user.uid}`))
-        const querySnapshot = await getDocs(q)
-        querySnapshot.forEach( doc => {
-            userDocs.push(doc.data())
-        })
-        console.log(userDocs)
+    //     const userDocs = []
+    //     const q = query(collection(db, 'HABC-responses'), where('createdBy', '==', `${user.uid}`))
+    //     const querySnapshot = await getDocs(q)
+    //     querySnapshot.forEach( doc => {
+    //         userDocs.push(doc.data())
+    //     })
+    //     console.log(userDocs)
 
 
-    }
+    // }
 
     // React.useEffect(() => {
     //     getResults()
@@ -27,7 +27,7 @@ export default function Results(){
 
     return(
         <>
-            <button onClick={getResults}>console log saved docs by user</button>
+            {/* <button onClick={getResults}>console log saved docs by user</button> */}
         </>
     )
 }
